Add interp1d test and last reorganise
